refactor: rename interfaces and update imports for Nostr integration
- Renamed interface file from interfaces.js to nostr-interface.js for clarity.
- Updated imports in marmot-client.ts, marmot-group.ts, and other related files to reflect the new interface structure.
- Adjusted type definitions in proposal files to use ProposalAction instead of ProposalBuilder for consistency.
