+++ src/client/group/proposals/invite-user.ts
@@ -2,12 +2,12 @@ import { isEvent, NostrEvent } from "applesauce-core/helpers/event";
 import { type KeyPackage } from "ts-mls";
 import { type ProposalAdd } from "ts-mls/proposal.js";
 import { getKeyPackage } from "../../../core/key-package.js";
-import { ProposalBuilder } from "../marmot-group.js";
+import { ProposalAction } from "../marmot-group.js";
 
 /** Builds a proposal to invite a user to the group from a key package event or raw key package */
 export function proposeInviteUser(
   keyPackageEvent: KeyPackage | NostrEvent,
-): ProposalBuilder<ProposalAdd> {
+): ProposalAction<ProposalAdd> {
   return async () => {
     const keyPackage = isEvent(keyPackageEvent)
       ? getKeyPackage(keyPackageEvent)
