+++ src/client/marmot-client.ts
@@ -18,7 +18,7 @@ import { createSimpleGroup, SimpleGroupOptions } from "../core/group.js";
 import { generateKeyPackage } from "../core/key-package.js";
 import { GroupStore } from "../store/group-store.js";
 import { KeyPackageStore } from "../store/key-package-store.js";
-import { NostrPool } from "./interfaces.js";
+import { NostrPool } from "./nostr-interface.js";
 import { MarmotGroup } from "./group/marmot-group.js";
 
 export type MarmotClientOptions = {
